+++ examples/controller.rs
@@ -1,0 +1,81 @@
+use bevy::prelude::*;
+use bevy_third_person_camera::{camera::*, controller::*, *};
+
+fn main() {
+    App::new()
+        .add_plugins((DefaultPlugins, ThirdPersonCameraPlugin))
+        .add_systems(Startup, (spawn_player, spawn_world, spawn_camera))
+        .run();
+}
+
+#[derive(Component)]
+struct Player;
+
+fn spawn_player(mut commands: Commands, assets: Res<AssetServer>) {
+    let player = (
+        SceneBundle {
+            scene: assets.load("Player.gltf#Scene0"),
+            transform: Transform::from_xyz(0.0, 0.5, 0.0),
+            ..default()
+        },
+        Player,
+        ThirdPersonCameraTarget,
+        ThirdPersonController::default(), // add third person controller
+    );
+
+    commands.spawn(player);
+}
+
+fn spawn_camera(mut commands: Commands) {
+    let camera = (
+        Camera3dBundle {
+            transform: Transform::from_xyz(-2.0, 2.5, 5.0).looking_at(Vec3::ZERO, Vec3::Y),
+            ..default()
+        },
+        ThirdPersonCamera {
+            aim_enabled: true,
+            aim_speed: 3.0, // default
+            aim_zoom: 0.7,  // default
+            offset_enabled: true,
+            offset_toggle_enabled: true,
+            gamepad_settings: CameraGamepadSettings { ..default() },
+            zoom_enabled: true,        // default
+            zoom: Zoom::new(1.5, 3.0), // default
+            ..default()
+        },
+    );
+    commands.spawn(camera);
+}
+
+fn spawn_world(
+    mut commands: Commands,
+    mut meshes: ResMut<Assets<Mesh>>,
+
+    mut materials: ResMut<Assets<StandardMaterial>>,
+) {
+    let floor = PbrBundle {
+        mesh: meshes.add(Mesh::from(Plane3d::default().mesh().size(15.0, 15.0))),
+        material: materials.add(Color::DARK_GREEN),
+        ..default()
+    };
+
+    let block = PbrBundle {
+        mesh: meshes.add(Mesh::from(Cuboid::new(0.5, 0.5, 0.5))),
+        transform: Transform::from_xyz(2.5, 0.25, -2.5),
+        material: materials.add(Color::BLUE),
+        ..default()
+    };
+
+    let light = PointLightBundle {
+        point_light: PointLight {
+            intensity: 1500.0 * 1000.0,
+            ..default()
+        },
+        transform: Transform::from_xyz(0.0, 5.0, 0.0),
+        ..default()
+    };
+
+    commands.spawn(block);
+    commands.spawn(floor);
+    commands.spawn(light);
+}